Added facet to anomalies plot
