+++ atmospheric_explorer/plotting/yearly_flux.py
@@ -3,13 +3,9 @@ APIs for generating dynamic and static plots
 """
 from __future__ import annotations
 
-import re
-from math import ceil
 from textwrap import dedent
 
 import geopandas as gpd
-import pandas as pd
-import plotly.express as px
 import plotly.graph_objects as go
 import xarray as xr
 
@@ -22,136 +18,9 @@ from atmospheric_explorer.data_transformations import (
     confidence_interval,
 )
 from atmospheric_explorer.loggers import get_logger
-from atmospheric_explorer.utils import hex_to_rgb
+from atmospheric_explorer.plotting.plot_utils import line_with_ci_subplots
 
 logger = get_logger("atmexp")
-
-
-def _add_ci(
-    fig: go.Figure,
-    trace: go.Figure,
-    dataset: pd.DataFrame,
-    labels: list[str],
-) -> None:
-    """Add confidence intervals to a plotly trace"""
-    line_color = ",".join([str(n) for n in hex_to_rgb(trace.line.color)])
-    if len(labels) > 1:
-        regex = re.compile("label=(.+?)(?=<br>)")
-        label = regex.search(trace.hovertemplate)[1]
-    else:
-        label = labels[0]
-    label_index = labels.index(label)
-    total_rows = ceil(len(labels) / 2)
-    label_row = total_rows - label_index // 2
-    label_col = label_index % 2 + 1
-    df_label = dataset[dataset["label"] == label]
-    times = df_label.index.tolist()
-    fig.add_trace(
-        go.Scatter(
-            x=times,
-            y=df_label["lower"].to_list(),
-            line_color=f"rgba({line_color}, 0)",
-            fill=None,
-            fillcolor=f"rgba({line_color}, 0.2)",
-            showlegend=False,
-            mode="lines",
-            name=f"CI {trace.legendgroup}",
-            legendgroup=f"CI {trace.legendgroup}",
-            hovertemplate="Lower: %{y}<extra></extra>",
-            hoverlabel={"bgcolor": f"rgba({line_color}, 0.2)"},
-        ),
-        row=label_row,
-        col=label_col,
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=times,
-            y=df_label["upper"].to_list(),
-            line_color=f"rgba({line_color}, 0)",
-            fill="tonexty",
-            fillcolor=f"rgba({line_color}, 0.2)",
-            showlegend=False,
-            mode="lines",
-            name=f"CI {trace.legendgroup}",
-            legendgroup=f"CI {trace.legendgroup}",
-            hoverinfo="name+text+y",
-            hovertemplate="Upper: %{y}<extra></extra>",
-            hoverlabel={"bgcolor": f"rgba({line_color}, 0.2)"},
-        ),
-        row=label_row,
-        col=label_col,
-    )
-
-
-def line_with_ci_subplots(
-    dataset: pd.DataFrame, col_num: int, unit: str, title: str
-) -> go.Figure:
-    """\
-    Facet line plot on countries/administrative entinties.
-    This function plots the yearly mean of a quantity along with its CI.
-
-    Parameters:
-        dataset (pd.DataFrame): pandas dataframe with (at least) columns
-                                    'label','color','mean','lower','upper'
-        col_num (int): number of maximum columns in the facet plot
-        unit (str): unit of measure
-        title (str): plot title
-    """
-    labels = sorted(dataset["label"].unique())
-    colors = sorted(dataset["color"].unique())
-    if len(labels) > 1:
-        fig = px.line(
-            data_frame=dataset,
-            y="value",
-            color="color",
-            facet_col="label",
-            facet_col_wrap=col_num,
-            facet_col_spacing=0.04,
-            facet_row_spacing=0.15 if len(labels) > 3 else 0.2,
-            color_discrete_sequence=px.colors.qualitative.D3,
-            category_orders={"color": colors, "label": labels},
-        )
-    else:
-        fig = px.line(
-            data_frame=dataset,
-            y="value",
-            color="color",
-            color_discrete_sequence=px.colors.qualitative.D3,
-            category_orders={"color": colors},
-        )
-    fig.for_each_trace(
-        lambda tr: _add_ci(
-            fig,
-            tr,
-            dataset[dataset["color"] == tr.legendgroup],
-            labels,
-        )
-    )
-    fig.for_each_annotation(
-        lambda a: a.update(text=a.text.split("label=")[-1], font={"size": 14})
-    )
-    fig.update_yaxes(title=unit, col=1)
-    fig.update_yaxes(showticklabels=True, matches=None)
-    fig.update_xaxes(showticklabels=True, matches=None)
-    total_rows = ceil(len(labels) / 2)
-    if len(labels) % 2 != 0:
-        fig.update_xaxes(title="Year", col=2, row=total_rows)
-    base_height = 220 if len(labels) >= 3 else 300
-    fig.update_layout(
-        title={
-            "text": title,
-            "x": 0.5,
-            "xanchor": "center",
-            "xref": "paper",
-            "font": {"size": 19},
-        },
-        height=base_height * total_rows,
-        hovermode="closest",
-    )
-    fig.update_traces(mode="lines+markers")
-    fig.update_traces(selector=-2 * len(labels) - 1, showlegend=True)  # legend for ci
-    fig.update_traces(selector=-1, showlegend=True)  # legend for ci
-    return fig
 
 
 def _ghg_flux_over_full_area(dataset: xr.Dataset, var_name: str):
@@ -304,11 +173,6 @@ def ghg_surface_satellite_yearly_plot(
     da_converted_agg = _ghg_surface_satellite_yearly_data(
         data_variable, years, months, var_name, shapes, add_satellite_observations
     )
-    # Clip countries
-    if shapes is not None:
-        col_num = 2 if len(shapes) >= 2 else 1
-    else:
-        col_num = 1
     # Pandas is easier to use for plotting
     df_pandas = (
         da_converted_agg.to_dataframe()
@@ -318,5 +182,5 @@ def ghg_surface_satellite_yearly_plot(
         .rename({"mean": "value"}, axis=1)
     )
     return line_with_ci_subplots(
-        df_pandas, col_num, da_converted_agg.attrs["units"], title
+        df_pandas, da_converted_agg.attrs["units"], title, add_ci=True
     )
